requirements.txt
streamlit==1.33.0
tensorflow==2.17.1
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
matplotlib==3.8.4
pandas==2.2.2
ipykernel==6.29.4

kaggle==1.5.13
opencv-python==4.7.0.72
tensorboard